Move string converters to utilities
